move everything away from helm struct
